+++ lib/pss/include/pss/comp/lex.h
@@ -84,7 +84,7 @@ typedef struct {
 	uint32_t    offset;                     /*!< the offset of this token in the source code*/
 	pss_comp_lex_token_type_t type;         /*!< the type of this token */
 	union {
-		int32_t i;                          /*!< the integer value of this token */
+		int64_t i;                          /*!< the integer value of this token */
 		char s[4096];                       /*!< the string value of this token */
 		const char* e;                      /*!< the error message */
 		pss_comp_lex_keyword_t k;           /*!< the keyword data */
